modified:   src/tmx/Asn_J2735/CMakeLists.txt 	modified:   src/tmx/TmxUtils/src/SNMPClient.cpp 	modified:   src/tmx/TmxUtils/src/SNMPClient.h 	modified:   src/tmx/tmx.doxyfile 	modified:   src/v2i-hub/ImmediateForwardPlugin/manifest.json 	modified:   src/v2i-hub/ImmediateForwardPlugin/src/ImmediateForwardPlugin.cpp 	modified:   src/v2i-hub/ImmediateForwardPlugin/src/ImmediateForwardPlugin.h

+++ src/tmx/TmxUtils/src/SNMPClient.cpp
@@ -4,7 +4,7 @@ namespace tmx::utils {
     
     // Client defaults to SNMPv3
     snmp_client::snmp_client(const std::string& ip, const int& port, const std::string& community, 
-        const std::string &snmp_user, const std::string &securityLevel, const std::string &authPassPhrase, int snmp_version, int timeout)
+        const std::string &snmp_user, const std::string &securityLevel, const std::string &authPassPhrase, uint8_t snmp_version, int timeout)
 
         : ip_(ip), port_(port), community_(community),snmp_version_(snmp_version), timeout_(timeout)
     {
@@ -14,7 +14,7 @@ namespace tmx::utils {
         PLOG(logINFO) << "Target device NTCIP port: " << port_;
 
         // Bring the IP address and port of the target SNMP device in the required form, which is "IPADDRESS:PORT":
-        std::string ip_port_string = ip_ + ":" + std::to_string(port_);    
+        std::string ip_port_string = ip_ + ":" + std::to_string(port_);
         char* ip_port = &ip_port_string[0];
         
         // Initialize SNMP session parameters
